Distributed Keystone for Distributed Cloud - Endpoint filters
Bringing in the Endpoint Filter Group blueprint into Openstack
client. Endpoint filter groups have been around in Keystone since the
days of KILO but only picked up by Openstack client in v3.15.

In the absence of Endpoint filter groups, clients are responsible for
filtering the endpoint list, instead of Keystone sending a tailored
list. For e.g. in Distributed Cloud, local Openstack services in the Central Region
should not be seeing endpoints for subclouds, as they don't directly
communicate with subclouds.

As part of this blueprint we also had to bring in an important
commit that Optimizes the command "openstack endpoint list". The current
behavior is to make a Service call for each Endpoint in the Endpoint
list, so for N endpoints yo have N + 1 calls to Keystone. The new
behavior makes 1 call for Endpoint list, and 1 call for Service list and
works with the two lists... so you do down from (N+1) calls to 2

Change-Id: I9c03938c25b56d64b59ce42cae5026f2830f02b7