Distributed Keystone for Distributed Cloud: Horizon
In Distributed Cloud, Keystone is now running on each Subcloud.
Switching to Subcloud region now requires Openstack Auth to retrieve an
Unscoped token from the switched Region and reinitialize the django
session and cookie data with token data retrieved from the Subcloud.

Since Subcloud's Keystone doesn't contain the Identity endpoint for the
Central Region, there was no way to go back in Horizon from a subcloud
region to the SystemController region. We achieve this by caching the
SystemController endpoint in the Django Session at the time of login.

Story: 2002842
Task: 22785

Change-Id: I274e8b0756e0f24321a108c6c1a0a5d6178e0c7a
Signed-off-by: Jack Ding <jack.ding@windriver.com>