move wsgi app/scriplets from stx-keystone to stx-upstream
In an effort to reduce the number of stx specific code
changes in the openstack core repos, this modification
moves some small python scriplets into stx-upstream and
adjust the spec file to pick the files from their new
location.

There is no changes to the content of the built rpm,
and no changes to the installed software.

This commit does not include modification to the repo
that contains the original sources, and does not require
the original sources to be modified or removed.
Removal of the original sources will be done
separately and can only be done once this commit gets
merged.

Story: 2003112
Task: 23231

Change-Id: I3e86e8e952adb447e76cf970d71b5b37fccc210d
Signed-off-by: Paul-Emile Element <Paul-Emile.Element@windriver.com>