unset horizon TIME_ZONE to use system value instead
The horizon configuration file has a value set for
TIME_ZONE (set to UTC). This value is not available for users to
modify and takes precedence over the system settings.
The result is that the horizon GUI will show times in UTC
regardless of the system settings

This commit simple unsets the value so that the GUI
uses the system values by default.

Closes-Bug: #1806051

Change-Id: If24daaa97109d1020ec7f4ebdafebb5cda395ccb
Signed-off-by: Paul-Emile Element <Paul-Emile.Element@windriver.com>