Increase keystone api probe and termination timeout, and clean up the stale files
The keystone-api pod gets stuck in CrashLoopBackOff on
AIO-SX lock/unlock attempt.

When Kubernetes decides to kill the keystone-api pod due to readiness
probe failure or other reasons, it calls the preStop hook immediately
before the container is terminated. This hook starts a graceful
shutdown process which includes removing pid, shared memory segment
and wsgi sock files. If the container is not terminated within the
grace period, a SIGKILL is sent, and the container is forced to shut
down.

It was observed that the keystone-api readiness probe was failing
frequently during booting. The failure caused subsequent shut
down/restart. In some cases, the container was forced to terminate
without cleanup, and the stale files were left behind.

On the restart, the application detected the file existed, and
treated it as configuration failure, hence the exit. As a result,
the pod went into a crash loop.

This update increases the readiness probe timeout and termination
grace period. It prevents the container to be restarted due to
slow initialization and gives the process more time to gracefully
shut down.

As the graceful shutdown is not guaranteed 100% of the time, this
update removes any stale files.

Story: 2004520
Task: 28392

Change-Id: I613a0db674de9578b3f9d1fa781a1612d9caf214
Signed-off-by: Tao Liu <tao.liu@windriver.com>