Fix the version string in cinder and glance clients
cinderclient was showing 0.0.0 for cinder --version
Same problem for glance.

The pbr version needed to be set when building from
outside of a git tree.

All other clients had this set properly.

This bug was introduced when the new stein clients were
added.

The cinderclient also needed some BuildRequires updated.

Closes-Bug: 1814573
Change-Id: I4afe783e25ab2172ae999787e6b0e3ec91f78419
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>