Remove problematic upstream mariadb commit
When removing the mariadb release from the cluster, this upstream commit
produces mariadb-ingress pods that are stuck in the "Terminating" state
with the associated containers becoming hung. This ultimately impacts
certain docker operations leading to PLEG health issues in the cluster.

The root cause of this is that the ingress pod uses dumb-init to start
the nginx-ingress-controller process. When the mariadb-ingress pod
terminates (via kill TERM -1) all child processes are terminated but the
docker-containerd-shim remains causing the hung container condition.

Temporarily reverting this commit. A fix will be introduced upstream
dealing with dumb-init and this commit will be pick up again on the next
full chart rebase.

Change-Id: I25ad9bc3213468a9060e741917d96d9ac5c01b40
Story: 2004520
Task: 29420
Signed-off-by: Robert Church <robert.church@windriver.com>