Update nova helm chart to fix console addressing
Upstream nova helm chart attempts to figure out the address for VM
consoles by running an init container that checks for ip routes and
addresses on a compute host.  It then sets the appropriate nova config
options in a config file which it passes to nova-compute.  However this
effectively overwrites the same config option that stx has already
set in nova.conf via per host overrides causing us to communicate over
the wrong network or not to connect at all.

This fix introduces an option to enable/disable passing of this
additional config file to nova-compute.  Default upstream behaviour is
unchanged, but we will disable it in our overrides.

Partial-Bug: #1815490
Change-Id: I86eb80578b23fd89b7f9643b943ee759f26a15be
Signed-off-by: Gerry Kopec <gerry.kopec@windriver.com>