Fix build break for python-openstackclient
A patch was being carried in stx-upstream for the client
that was meant to be removed once the patch was merged
in the python-openstackclient git.

It merged, which meant attempts to build would fail as the
patch was being applied twice.

This fix is to remove the temporary patch.

Closed-Bug: 1819176
Change-Id: Ib638367361d411ccbca03e85e30e0b2931738956
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>