Remove most openstack components from ISO
Most of the openstack components are containerized
except for keystone, barbican, horizon and the clients.

This commit removes the other RPMs from the ISO.
If RPMs are still in use through RPM requirements,
they will continue to be pulled into the ISO until
those spec files are updated.

Story: 2004764
Task: 30105
Depends-On: Ifa0171b06e23fd77d373983d644df3f56ae4e2de
Change-Id: I4109398df9ff9f5608721f70611b7554fc31eacb
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>