Explicitly include heatclient in the ISO
Several openstack clients are not explicitly included
in the ISO, but are being implicitly included.

This commit explicitly adds those clients, so that
if the components that are implicitly including them
are removed, the clients will still remain.

heatclient was the only component in this update that
was missing.

A later commit to build-tools/build_iso/image.inc
should be submitted to clean its entries.

Closes-Bug: 1822200
Change-Id: Idb11e6217df1f30725f357fdfdc3864a6d33474a
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>